+++ src/bzflag/HUDuiControl.cxx
@@ -103,7 +103,7 @@ void HUDuiControl::setParent(HUDuiControl* parentControl)
   parent = parentControl;
 }
 
-bool HUDuiControl::isAtNavQueueIndex(int index)
+bool HUDuiControl::isAtNavQueueIndex(size_t index)
 {
   if (!navList) return false;
   return navList->at(index) == this;
